requirements.txt
requests==2.18.4
python-whois
